New rule: don't rewrite paths unless src_path is defined
This clarifies the difference between files that are being
exported from the source filesystem, and other files which
might be programmatically generated.  The latter should be
using the known paths of the dst rather than expecting to
get rewritter.

Also implemented more of the interpreter rewriting for ELF
and scripts.