Translate user->expire field on FreeBSD/linux - #1
In case perl's pw_has ever returns true for the 'expire' field,
pass it through _days_since_1970_to_time, and on BSD store the
expire field into expire_time since it is already the correct
format.

Meanwhile, also list BSD's 'class' as a known attr of User objects.
This was generating warnings in the tests.