Restructure ::Unix->add to share source-rewriting code
I wanted to be able to pass stat results of the src directory to ->add
without triggering a second call to lstat(), but hashrefs are implied
to be in terms of the destination rather than the source.  I decided
that if 'name' is missing and 'src_path' is not, that can be used as a
flag to indicate that source-translation code should be run for this
hashref.