Fix CPIO to correctly write file trailer on ->finish
I didn't read the spec closely enough, and missed that the file is
supposed to end with a "TRAILER!!!" entry.  I already had a dummy
->finish method, so I just upgraded that to actually do something.

I also added an 'autoclose' attribute, set to true by default if
the module also opened the file.  Also calling ->flush is fairly
important even if I don't close the file, because I was using
short ->print calls to pad to multiple of 4 and those might not
get flushed automatically.