Fix bugs where src_abs_path failed when src was '/'
Embarrassing, because src='/' is the main use case for the module.
It's just a bit hard to test since I can't know anything about the
host it is running on.  One thing I do know is that the "$tmp" path
can be reached from "/", so I was able to make that into a test.