New _trace_deps feature uses strace to observe opened files
The best automated way to collect dependencies of a tool is to
execute it and use strace to observe every file it opens.  While
this only works in a very limited case of using a Linux host with
strace installed to export a Linux image, that's probably the most
likely use case.