Simplify VFAT directory entries to hashes instead of objects
I decided to store file attributes on the File/Directory objects
and let the directory entry optionally override them.  Then, the
directory entry really only needs the long and short name, and a
reference to a file.  This makes them simple enough that they
don't really need an object, which makes the code simpler in some
places.

I also moved the overwrite-attributes-of-an-autovivified-dir logic
to the VFAT->add method instead of the Dir->add method, which seems
to clean things up in the Directory class.

I also fixed bugs in the is_valid_longname and is_valid_shortname
exports, and made build_shortname an export.