LazyFileData, and eliminate data_path parameter from %fileinfo
I decided I don't like having to check for ->{data} and also
->{data_path} in so many places, especially since they could be
contradictory.  I made map_or_load_file public, and then also
wrapped it with a new object LazyFileData which can carry the
filename to the destination or magically  load it on demand.
All the ::Unix and ::CPIO methods now support padding the ->{data}
as a scalar ref, a plain scalar, or a LazyFileData object.