Refactor VFAT so that Directories are separate from Files
I decided I could best re-use the Directory objects in ISO9660 if
they don't have FAT attributes on them.  Also, if I ever wanted to
load a VFAT filesystem it would help if I could create the File
objects first and then inflate the Directory objects afterward.

So, now a directory entry is a very minimal name/shortname/file
and if it is a directory it also gets a ->{dir} reference.
Autovivified directories are represented a simply lacking a
->{file} entry, which made things cleaner.  They get populated
while calculating the size of the directory.

Also, now all the packing routines are in the main VFAT module
where they belong.