Better label validation, and ability to override default time
I added a field 'default_time' which is used instead of repeatedly
calling time().  This allows the timestamps to be set to a known
value for reproducible builds.  Then I reviewed the character
restrictions on the label fields of the volume descriptor and more
accurately validate them, and also support passing file references
as the publisher/preparer/application labels.